Created .env environment configuration file
This .env file specifies the necessary Flask environment variables to run the scraping server with in conjunction with Michael's updates in the flask-cors-fix branch. A user must edit line 3 with the confidential Google OAuth Client ID before running the server with Flask (which is reflected in the edited README.md on this branch).